refactor: seperate data types on public and profile
