+++ src/components/DocumentsSection/DocumentsSection.tsx
@@ -1,7 +1,7 @@
 import cl from 'classnames';
 import { FC } from 'react';
 import { v4 as uuidv4 } from 'uuid';
-import { TDocument } from '../../types/data';
+import { TDocument } from '../../types/profileData';
 import { DocumentsCategory } from './DocumentsCategory/DocumentsCategory';
 import s from './DocumentsSection.module.scss';
