+++ src/components/LeadSection/LeadSection.tsx
@@ -1,7 +1,7 @@
 import cl from 'classnames';
 import { FC } from 'react';
 import { v4 as uuidv4 } from 'uuid';
-import { TProperty } from '../../types/data';
+import { TProperty } from '../../types/publicData';
 import { LeadSectionCards } from './LeadSectionCards/LeadSectionCards';
 import s from './LeadSection.module.scss';
