+++ src/components/LeadSection/LeadSectionCards/LeadSectionCards.tsx
@@ -2,7 +2,7 @@ import cl from 'classnames';
 import { FC } from 'react';
 import { Link } from 'react-router-dom';
 import { v4 as uuidv4 } from 'uuid';
-import { TProperty } from '../../../types/data';
+import { TProperty } from '../../../types/publicData';
 import s from './LeadSectionCards.module.scss';
 
 type IProps = {
