+++ src/components/ProductSliderSection/ProductSliderSection.tsx
@@ -4,7 +4,7 @@ import { v4 as uuidv4 } from 'uuid';
 import brokenScreen from '../../assets/svg/broken_screen.svg';
 import LapTop from '../../assets/svg/laptop.svg?svgr';
 import { API_BASE_URL } from '../../constants/externalLinks';
-import { TProductImg } from '../../types/data';
+import { TProductImg } from '../../types/publicData';
 import { SectionTitle } from '../SectionTitle/SectionTitle';
 import { ArrowInCircleIcon, SpinnerIcon } from '../ui/icons';
 import s from './ProductSliderSection.module.scss';
