+++ src/components/ProfileNav/ProfileProjectsNav/ProfileProjectsNav.tsx
@@ -2,7 +2,7 @@ import cl from 'classnames';
 import { FC } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import { ROUTE_PROFILE_PROJECTS } from '../../../constants/routesConstants';
-import { TProjectShortInfo } from '../../../types/data';
+import { TProjectShortInfo } from '../../../types/profileData';
 import { ArrowFatRightIcon } from '../../ui/icons';
 import { ProfileNavLink } from '../ProfileNavLink/ProfileNavLink';
 import s from './ProfileProjectsNav.module.scss';
