+++ src/components/ProjectCard/ProjectCard.tsx
@@ -1,10 +1,10 @@
 import cl from 'classnames';
 import { FC } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { Text } from '../ui/Text/Text';
 import { ROUTE_PROFILE_PROJECTS } from '../../constants/routesConstants';
-import { TProjectShortInfo } from '../../types/data';
+import { TProjectShortInfo } from '../../types/profileData';
 import { convertDateToShortFormat } from '../../utils/dateConvertHelpers';
+import { Text } from '../ui/Text/Text';
 import { ProgressCircle } from './ProgressCircle/ProgressCircle';
 import s from './ProjectCard.module.scss';
