+++ src/components/TeamSliderSection/TeamSliderSection.tsx
@@ -5,7 +5,7 @@ import 'react-multi-carousel/lib/styles.css';
 import { CarouselProps, ResponsiveType } from 'react-multi-carousel/lib/types';
 import { useMediaQuery } from 'react-responsive';
 import { v4 as uuidv4 } from 'uuid';
-import { TEmployee } from '../../types/data';
+import { TEmployee } from '../../types/publicData';
 import { SectionTitle } from '../SectionTitle/SectionTitle';
 import { CarouselBtn } from './CarouselBtn/CarouselBtn';
 import { TeamGradientSlide } from './TeamGradientSlide/TeamGradientSlide';
